# Store Prometheus labels as a slice instead of a map