--- a/cmd/otel-allocator/benchmark_test.go
+++ b/cmd/otel-allocator/benchmark_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/discovery"
 	"github.com/prometheus/prometheus/discovery/targetgroup"
+	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/model/relabel"
 	"github.com/stretchr/testify/require"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -48,13 +49,14 @@ func BenchmarkProcessTargets(b *testing.B) {
 	targetsPerGroup := 5
 	groupsPerJob := 20
 	tsets := prepareBenchmarkData(numTargets, targetsPerGroup, groupsPerJob)
+	labelsBuilder := labels.NewBuilder(labels.EmptyLabels())
 
 	b.ResetTimer()
 	for _, strategy := range allocation.GetRegisteredAllocatorNames() {
 		b.Run(strategy, func(b *testing.B) {
 			targetDiscoverer, allocator := createTestDiscoverer(strategy, map[string][]*relabel.Config{})
 			for i := 0; i < b.N; i++ {
-				targetDiscoverer.ProcessTargets(tsets, allocator.SetTargets)
+				targetDiscoverer.ProcessTargets(labelsBuilder, tsets, allocator.SetTargets)
 			}
 		})
 	}
@@ -67,6 +69,7 @@ func BenchmarkProcessTargetsWithRelabelConfig(b *testing.B) {
 	targetsPerGroup := 5
 	groupsPerJob := 20
 	tsets := prepareBenchmarkData(numTargets, targetsPerGroup, groupsPerJob)
+	labelsBuilder := labels.NewBuilder(labels.EmptyLabels())
 	prehookConfig := make(map[string][]*relabel.Config, len(tsets))
 	for jobName := range tsets {
 		// keep all targets in half the jobs, drop the rest
@@ -93,7 +96,7 @@ func BenchmarkProcessTargetsWithRelabelConfig(b *testing.B) {
 		b.Run(strategy, func(b *testing.B) {
 			targetDiscoverer, allocator := createTestDiscoverer(strategy, prehookConfig)
 			for i := 0; i < b.N; i++ {
-				targetDiscoverer.ProcessTargets(tsets, allocator.SetTargets)
+				targetDiscoverer.ProcessTargets(labelsBuilder, tsets, allocator.SetTargets)
 			}
 		})
 	}