--- a/cmd/otel-allocator/prehook/relabel.go
+++ b/cmd/otel-allocator/prehook/relabel.go
@@ -43,17 +43,14 @@ func (tf *RelabelConfigTargetFilter) Apply(targets map[string]*target.Item) map[
 
 	// Note: jobNameKey != tItem.JobName (jobNameKey is hashed)
 	for jobNameKey, tItem := range targets {
-		keepTarget := true
+		var keepTarget bool
 		lset := tItem.Labels
 		for _, cfg := range tf.relabelCfg[tItem.JobName] {
 			lset, keepTarget = relabel.Process(lset, cfg)
 			if !keepTarget {
+				delete(targets, jobNameKey)
 				break // inner loop
 			}
-		}
-
-		if !keepTarget {
-			delete(targets, jobNameKey)
 		}
 	}